Fix back-referencing problem with GuardedData serialization in CC
CC has a custom provider serialization logic that doesn't play well with
owner references of the guarded data. By default, owner is serialized as
every other provider - computing execution-time-value, etc. This isn't
suitable for the GuardedData, because the owner's identity must be
preserved.

Moreover, the serialization code assumes that the serialization of the
owner has already completed. This holds for cases where guarded data is
the field of the owner, as it should be.